changing API so that user can provide `size_t` length rather than `unsigned` length

+++ murmur3.h
@@ -8,8 +8,8 @@
 #ifndef MURMUR3_H
 #define MURMUR3_H
 
-void MurmurHash3_x86_32 (const void *key, unsigned len, unsigned seed, void *out);
-void MurmurHash3_x86_128(const void *key, unsigned len, unsigned seed, void *out);
-void MurmurHash3_x64_128(const void *key, unsigned len, unsigned seed, void *out);
+void MurmurHash3_x86_32 (const void *key, size_t len, unsigned seed, void *out);
+void MurmurHash3_x86_128(const void *key, size_t len, unsigned seed, void *out);
+void MurmurHash3_x64_128(const void *key, size_t len, unsigned seed, void *out);
 
 #endif /* MURMUR3_H */
